Corrected gamification
Corrected some small smelling errors and added some comments. I am not
sure if the section include enough analysis or if the purpose of the
section is analysis or description of gamification. Some analysis about
techneques to improve game experience is included, which is good and the
section does include some analysis, which is also good.